@@ -8,12 +8,11 @@ readme = "README.md"
 packages = [{include = "api_tabular"}]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.11"
 aiohttp = "^3.8.4"
 aiohttp-cors = "0.7.0"
 aiohttp-swagger = "1.0.16"
 sentry-sdk = "^1.25.1"
-toml = "^0.10.2"
 
 [tool.poetry.group.dev.dependencies]
 aiohttp-devtools = "^1.0.post0"
