@@ -1,5 +1,4 @@
 import json
-from typing import Union
 
 import sentry_sdk
 from aiohttp import web
@@ -16,7 +15,7 @@ class QueryException(web.HTTPException):
         super().__init__(content_type="application/json", text=json.dumps(error_body))
 
 
-def handle_exception(status: int, title: str, detail: Union[str, dict], resource_id: str = None):
+def handle_exception(status: int, title: str, detail: str | dict, resource_id: str = None):
     event_id = None
     e = Exception(detail)
     if config.SENTRY_DSN:
